requirements.txt
# Core dependencies
streamlit>=1.27.0
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Spatial transcriptomics packages
scanpy>=1.9.0
anndata>=0.8.0

# AWS integration
boto3>=1.26.0
python-dotenv>=0.19.0

# Visualization
plotly>=5.10.0

# Other utilities
pyyaml>=6.0
requests>=2.28.0
pillow>=9.2.0
scipy>=1.7.0